refactor(generator_to_phpunit_data_provider): Move the function to a tested utility class
