+++ tests/phpunit/AutoReview/ProjectCode/ProjectCodeProvider.php
@@ -87,7 +87,7 @@ use Infection\Testing\SingletonContainer;
 use Infection\Testing\SourceTestClassNameScheme;
 use Infection\Testing\StringNormalizer;
 use Infection\Tests\AutoReview\ConcreteClassReflector;
-use function Infection\Tests\generator_to_phpunit_data_provider;
+use Infection\Tests\TestingUtility\PHPUnit\DataProviderFactory;
 use function iterator_to_array;
 use function ltrim;
 use function Pipeline\take;
@@ -218,7 +218,7 @@ final class ProjectCodeProvider
 
     public static function sourceClassesProvider(): iterable
     {
-        yield from generator_to_phpunit_data_provider(
+        yield from DataProviderFactory::fromIterable(
             self::provideSourceClasses(),
         );
     }
@@ -233,7 +233,7 @@ final class ProjectCodeProvider
 
     public static function concreteSourceClassesProvider(): iterable
     {
-        yield from generator_to_phpunit_data_provider(
+        yield from DataProviderFactory::fromIterable(
             self::provideConcreteSourceClasses(),
         );
     }
@@ -274,7 +274,7 @@ final class ProjectCodeProvider
 
     public static function sourceClassesToCheckForPublicPropertiesProvider(): iterable
     {
-        yield from generator_to_phpunit_data_provider(
+        yield from DataProviderFactory::fromIterable(
             self::provideSourceClassesToCheckForPublicProperties(),
         );
     }
@@ -315,14 +315,14 @@ final class ProjectCodeProvider
     // test instead of a test provider.
     public static function classesTestProvider(): iterable
     {
-        yield from generator_to_phpunit_data_provider(
+        yield from DataProviderFactory::fromIterable(
             self::provideTestClasses(),
         );
     }
 
     public static function nonTestedConcreteClassesProvider(): iterable
     {
-        yield from generator_to_phpunit_data_provider([
+        yield from DataProviderFactory::fromIterable([
             ...self::NON_TESTED_CONCRETE_CLASSES,
             ...self::CONCRETE_CLASSES_WITH_TESTS_IN_DIFFERENT_LOCATION,
         ]);
@@ -330,7 +330,7 @@ final class ProjectCodeProvider
 
     public static function nonFinalExtensionClasses(): iterable
     {
-        yield from generator_to_phpunit_data_provider(
+        yield from DataProviderFactory::fromIterable(
             self::NON_FINAL_EXTENSION_CLASSES,
         );
     }
